[core] Move `helper` types into `@mui/x-internals` package
