--- packages/x-charts/src/internals/SlotComponentPropsFromProps.ts
@@ -1,5 +1,0 @@
-export type SlotComponentPropsFromProps<
-  TProps extends {},
-  TOverrides extends {},
-  TOwnerState extends {},
-> = (Partial<TProps> & TOverrides) | ((ownerState: TOwnerState) => Partial<TProps> & TOverrides);
