+++ packages/x-tree-view/src/internals/models/helpers.ts
@@ -1,19 +1,5 @@
 import type { TreeViewAnyPluginSignature, TreeViewPlugin } from './plugin';
 
-export type DefaultizedProps<
-  P extends {},
-  RequiredProps extends keyof P,
-  AdditionalProps extends {} = {},
-> = Omit<P, RequiredProps | keyof AdditionalProps> &
-  Required<Pick<P, RequiredProps>> &
-  AdditionalProps;
-
-export type SlotComponentPropsFromProps<
-  TProps extends {},
-  TOverrides extends {},
-  TOwnerState extends {},
-> = (Partial<TProps> & TOverrides) | ((ownerState: TOwnerState) => Partial<TProps> & TOverrides);
-
 type IsAny<T> = 0 extends 1 & T ? true : false;
 
 export type OptionalIfEmpty<A extends string, B> = keyof B extends never
